warden routes were protected

+++ src/main/java/com/myapp/hostel_management_system/controller/StudentController.java
@@ -1,6 +1,7 @@
 package com.myapp.hostel_management_system.controller;
 
 import com.myapp.hostel_management_system.entity.Student;
+import com.myapp.hostel_management_system.service.AuthorizeService;
 import com.myapp.hostel_management_system.service.HostelService;
 import com.myapp.hostel_management_system.service.StudentService;
 import org.springframework.stereotype.Controller;
@@ -17,61 +18,83 @@ public class StudentController {
     private final StudentService studentService;
     private final HostelService hostelService;
 
-    public StudentController(StudentService studentService, HostelService hostelService) {
-        super();
+    private final AuthorizeService authorizeService;
+
+    public StudentController(StudentService studentService, HostelService hostelService, AuthorizeService authorizeService) {
         this.studentService = studentService;
         this.hostelService = hostelService;
+        this.authorizeService = authorizeService;
     }
 
     @GetMapping("/admin-students")
     public String listStudents(Model model) {
-        model.addAttribute("students", studentService.getAllStudents());
-
-        return "app/admin/student/index";
+        if (authorizeService.warden()) {
+            model.addAttribute("students", studentService.getAllStudents());
+            return "app/admin/student/index";
+        } else {
+            return "redirect:/";
+        }
     }
 
     @GetMapping("/admin-students/add")
     public String createStudents(Model model) {
-        Student student = new Student();
-        model.addAttribute("students", studentService.getAllStudents());
-
-        return "app/admin/student/add";
+        if (authorizeService.warden()) {
+            Student student = new Student();
+            model.addAttribute("students", studentService.getAllStudents());
+            return "app/admin/student/add";
+        } else {
+            return "redirect:/";
+        }
     }
 
     @PostMapping("/admin-students")
     public String saveStudent(@ModelAttribute("students") Student student, BindingResult result) {
-        if (student.getPassword() == null) {
-            student.setPassword("abcd1234");
+        if (authorizeService.warden()) {
+            if (student.getPassword() == null) {
+                student.setPassword("abcd1234");
+            }
+            studentService.studentSave(student);
+            return "redirect:/admin-students";
+        } else {
+            return "redirect:/";
         }
-        studentService.studentSave(student);
-        return "redirect:/admin-students";
     }
 
     @GetMapping("/admin-students/edit/{id}")
     public String editStudent(@PathVariable String id, Model model) {
-        model.addAttribute("student", studentService.getStudentById(id));
-        model.addAttribute("hostels", hostelService.getAllHostel());
-        return "app/admin/student/edit";
+        if (authorizeService.warden()) {
+            model.addAttribute("student", studentService.getStudentById(id));
+            model.addAttribute("hostels", hostelService.getAllHostel());
+            return "app/admin/student/edit";
+        } else {
+            return "redirect:/";
+        }
     }
 
     @PostMapping("/admin-students/{id}")
     public String updateStudent(@PathVariable String id,
                                 @ModelAttribute("students") Student student) {
-        Student existingStudent = studentService.getStudentById(id);
-        existingStudent.setId(id);
-        existingStudent.setFirstname(student.getFirstname());
-        existingStudent.setLastname(student.getLastname());
-        existingStudent.setEmail(student.getEmail());
-        existingStudent.setHostel(student.getHostel());
-        studentService.updateStudent(existingStudent);
-        return "redirect:/admin-students";
+        if (authorizeService.warden()) {
+            Student existingStudent = studentService.getStudentById(id);
+            existingStudent.setId(id);
+            existingStudent.setFirstname(student.getFirstname());
+            existingStudent.setLastname(student.getLastname());
+            existingStudent.setEmail(student.getEmail());
+            existingStudent.setHostel(student.getHostel());
+            studentService.updateStudent(existingStudent);
+            return "redirect:/admin-students";
+        } else {
+            return "redirect:/";
+        }
     }
 
     @GetMapping("/admin-students/{id}")
     public String deleteStudent(@PathVariable String id) {
-        studentService.deleteStudentById(id);
-        return "redirect:/admin-students";
-
+        if (authorizeService.warden()) {
+            studentService.deleteStudentById(id);
+            return "redirect:/admin-students";
+        } else {
+            return "redirect:/";
+        }
     }
-
 }
